+++ src/features/create-patient/create-patient.tsx
@@ -26,7 +26,7 @@ const CreatePatient = () => {
   const breadcrumbs = [
     {
         name: 'Inicio',
-        link: '/',
+        link: Routing.HOME,
         clickable: true,
         actual: false,
       },
