+++ src/features/obtain-model/obtain-model.tsx
@@ -24,13 +24,13 @@ const ObtainModelDrug = (props: any) => {
     const breadcrumbs = [
         {
             name: 'Inicio',
-            link: '/',
+            link: Routing.HOME,
             clickable: true,
             actual: false,
           },
           {
             name: 'Seleccionar modelo/paciente',
-            link: Routing.SELECCIONAR_PACIENTE,
+            link: Routing.SELECT_PATIENT_MODEL,
             clickable: true,
             actual: false,
           },
