+++ src/features/patients-page/patients-page.tsx
@@ -10,7 +10,7 @@ const PatientsPage = () => {
   const breadcrumbs = [
     {
         name: 'Inicio',
-        link: '/',
+        link: Routing.HOME,
         clickable: true,
         actual: false,
       },
